Unify stats views + deterministic column_id resolution
This change refactors connector statistics production to use a shared
ConnectorStatsViewBuilder and introduces deterministic column_id
computation via ColumnIdAlgorithm (Iceberg: CID_FIELD_ID,
Delta: CID_PATH_ORDINAL).

+++ core/catalog/src/main/java/ai/floedb/floecat/systemcatalog/spi/scanner/StatsProvider.java
@@ -26,7 +26,7 @@ public interface StatsProvider {
     return Optional.empty();
   }
 
-  default Optional<ColumnStatsView> columnStats(ResourceId tableId, int columnId) {
+  default Optional<ColumnStatsView> columnStats(ResourceId tableId, long columnId) {
     return Optional.empty();
   }
 
@@ -45,7 +45,7 @@ public interface StatsProvider {
   interface ColumnStatsView {
     ResourceId tableId();
 
-    int columnId();
+    long columnId();
 
     String columnName();
