+++ service/src/main/java/ai/floedb/floecat/service/repo/impl/StatsRepository.java
@@ -95,7 +95,7 @@ public class StatsRepository {
     columnStatsRepo.create(value);
   }
 
-  public Optional<ColumnStats> getColumnStats(ResourceId tableId, long snapshotId, int columnId) {
+  public Optional<ColumnStats> getColumnStats(ResourceId tableId, long snapshotId, long columnId) {
     return columnStatsRepo.getByKey(
         new ColumnStatsKey(tableId.getAccountId(), tableId.getId(), snapshotId, columnId, ""));
   }
@@ -237,19 +237,19 @@ public class StatsRepository {
         new TableStatsKey(tableId.getAccountId(), tableId.getId(), snapshotId, ""));
   }
 
-  public MutationMeta metaForColumnStats(ResourceId tableId, long snapshotId, int columnId) {
+  public MutationMeta metaForColumnStats(ResourceId tableId, long snapshotId, long columnId) {
     return columnStatsRepo.metaFor(
         new ColumnStatsKey(tableId.getAccountId(), tableId.getId(), snapshotId, columnId, ""));
   }
 
   public MutationMeta metaForColumnStats(
-      ResourceId tableId, long snapshotId, int columnId, Timestamp nowTs) {
+      ResourceId tableId, long snapshotId, long columnId, Timestamp nowTs) {
     return columnStatsRepo.metaFor(
         new ColumnStatsKey(tableId.getAccountId(), tableId.getId(), snapshotId, columnId, ""),
         nowTs);
   }
 
-  public MutationMeta metaForColumnStatsSafe(ResourceId tableId, long snapshotId, int columnId) {
+  public MutationMeta metaForColumnStatsSafe(ResourceId tableId, long snapshotId, long columnId) {
     return columnStatsRepo.metaForSafe(
         new ColumnStatsKey(tableId.getAccountId(), tableId.getId(), snapshotId, columnId, ""));
   }
